One function to perform integration
Reorganization of the code to avoid duplication.  `dointegrate' is the
function that actually performs integration, `Vegas', `Suave',
`Divonne', and `Cuhre' are wrappers around it.

In addition fix type of `nregions' variables, it is Cint but was defined
as Cdouble.  This was probably the caused of eccessive memory usage in
`Divonne'.

In test/benchmark.jl, recompile C and Fortran binaries only if
necessary.