Rename integrator functions and some keywords
All these changes aim at increasing consistency with the Julia
environment.  This is a Julia wrapper of an external library (Cuba), it
should target Julia users rather than Cuba users.

In Julia, uppercase functions are usually reserved to type constructor,
generic functions should be lowercase.

In addition, other integrator functions (like `quadgk' in Base and
functions in Cubature package) adopt "abstol" and "reltol" for absolute
and relative errors (like in Matlab), and "maxevals" for the maximum
number of evaluations.